move auditDiff enum
