+++ x-pack/plugins/audit/common/types.ts
@@ -21,10 +21,3 @@ export interface AuditLog {
   subjectId: AuditLogSchemaType['subjectId'];
   data: AuditLogSchemaType['data'];
 }
-
-export enum AuditDiffOperation {
-  ADD = 'add',
-  DELETE = 'delete',
-  UPDATE = 'update',
-  MOVE = 'move',
-}
